Move trap handler to its own file and make the additional trap types seeded
