+++ soh/soh/Enhancements/game-interactor/GameInteractor.h
@@ -389,6 +389,7 @@ typedef enum {
 
     /*** Give Items ***/
 
+    VB_SHORT_CIRCUIT_GIVE_ITEM_PROCESS,
     VB_FREEZE_ON_SKULL_TOKEN,
     // Opt: *EnBox
     VB_GIVE_ITEM_FROM_CHEST,
